requirements.txt
pyTelegramBotAPI==4.7.1
Flask==2.1.1
pymongo==3.12.0
python-dotenv==0.19.0
requests==2.25.1
